refactor: 폴더명 변경 userExample->userInfo, userIdSelector 생성
