+++ src/pages/DetailPage/DetailPage.tsx
@@ -24,7 +24,7 @@ import noImage from 'assets/images/NoImage.png';
 import noUser from 'assets/images/NoUser.gif';
 import timeAgo from 'utils/timeAgo';
 import { useSetRecoilState, useRecoilValue } from 'recoil';
-import { userProfileSelector } from 'recoil/userExample';
+import { userProfileSelector } from 'recoil/userInfo';
 import { useVerifyUser } from 'hooks';
 import { DetailMap } from 'components/detailPage';
 import { commentCountAtom } from 'recoil/commentCount/commentCountAtom';
