+++ src/recoil/userInfo/selector.ts
@@ -12,6 +12,14 @@ export const userProfileSelector = selector<UserState>({
   },
 });
 
+export const userIdSelector = selector<string>({
+  key: 'userIdSelector',
+  get: ({ get }) => {
+    const user = get(userState);
+    return user.userId;
+  },
+});
+
 //userIsDefinedSelector는 userState의 userId가 있는지 없는지를 boolean으로 반환합니다.
 export const userIsDefinedSelector = selector<boolean>({
   key: 'userIsDefinedSelector',
